Rename AT3 Folder, create AT3RadarTargetDisplay class
