+++ AT3/AT3RadarTargetDisplay.cpp
@@ -1,0 +1,16 @@
+#include "stdafx.h"
+#include "HKCPDisplay.hpp"
+#include "Constant.hpp"
+#include "EuroScopePlugIn.h"
+#include "AT3RadarTargetDisplay.hpp"
+
+using namespace EuroScopePlugIn;
+
+AT3RadarTargetDisplay::AT3RadarTargetDisplay() {
+
+}
+
+void AT3RadarTargetDisplay::OnRefresh(HDC hDC, int Phase, HKCPDisplay* Display)
+{
+
+}
