+++ AT3RadarTargetDisplay.cpp
@@ -1,0 +1,12 @@
+#include "stdafx.h"
+#include "HKCPDisplay.hpp"
+#include "Constant.hpp"
+#include "EuroScopePlugIn.h"
+#include "AT3Ra"
+
+using namespace EuroScopePlugIn;
+
+void AT3RadarTargetDisplay::OnRefresh(HDC hDC, int Phase, HKCPDisplay* Display)
+{
+
+}
