+++ HKCP.vcxproj
@@ -19,7 +19,8 @@
     </ProjectConfiguration>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="AT3Tags\AT3Tags.hpp" />
+    <ClInclude Include="AT3\AT3RadarTargetDisplay.hpp" />
+    <ClInclude Include="AT3\AT3Tags.hpp" />
     <ClInclude Include="Atis\AtisDisplay.hpp" />
     <ClInclude Include="Atis\AtisPlugin.hpp" />
     <ClInclude Include="Constant.hpp" />
@@ -32,7 +33,8 @@
     <ClInclude Include="VFPC\analyzeFP.hpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="AT3Tags\AT3Tags.cpp" />
+    <ClCompile Include="AT3\AT3RadarTargetDisplay.cpp" />
+    <ClCompile Include="AT3\AT3Tags.cpp" />
     <ClCompile Include="Atis\AtisDisplay.cpp" />
     <ClCompile Include="Atis\AtisPlugin.cpp" />
     <ClCompile Include="HKCP.cpp" />
